--- Runtime/CharacterLoader/CharacterManager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: e79db5b052a895b47a3e5290ea568dfb
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
