--- Tests/Loader/CharacterLoaderTests.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 56be65990a5cfd24dbc789f699b1e1e3
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
